added program controller and hospital program class
I've added the ProgramController routes with 2 get endpoints. GET your favorite doctor's program and GET a doctor's program by id. I have also added the hospital program class with two methods meant for verifying if the hours that a patient sends us is valid and a method to see the availabilty of a doctor.